Show incomplete banner if falling back to English
The current (without this PR) behaviour shows the incomplete banner when
there is an alt marked as incomplete, but not when the alt is completely
absent (Lektor fallbacks to the default contents.lr in this case). This
adds a custom filter in Jinja2 to detect this case, and shows the
incomplete banner.