templates: Use forkawesome as source for icons
This is a possible resolution to the issue of a "exposed fontawesome pro
link" mentioned by discord user "Lord Alfred".

forkawesome is a fork of fontawesome that is completely free for
commercial use. For full details, see
https://forkaweso.me/Fork-Awesome/license/

forkawesome is forked from a prior version of fontawesome at a time when
it was liberally licensed.

forkawesome doesn't have `fa-brands`, but discord & mastodon icons
are available in the base namespace. I verified locally that these icons
remain visible after my changes.

I did a quick search of all used fa- icons and I believe they are all
represented in forkawesome. Additionally the fa-lg class is also supported.