Don't Translate Event Names
Don't put event names in the po file when we're going to a conference.  Devs usually use english conference names when talking about things.  For example, https://tw.pycon.org/2025/zh-hant mentions the conference as PyCon Taiwan even in the Chinese Traditional text (which I can read a bit of).