Toga: "How to Help"
Adds a "How to Help" section.

Easter egg: the *first draft* of this section is justified on the right plus or minus 2 characters.

```
You can help by fixing bugs in or adding new features to existing widgets,
and implementing widgets on platforms for which they aren't supported yet.
If you have experience in, or would like to learn about the native GUI APIs
of a platform, you can start contributing to the widgets on that platform.
If you'd like to learn about Textual and terminal applications, you could
look into the Textual backend. If you posess web skills and would like to
learn more about Shoelace or PyScript, you could start contributing to the
Web backend, which needs a lot of help implementing new widgets. You could
also propose new widgets and implement it for at least one platform. If you
are *really* daring, you could implement a new platform backend -- we'd be
interested in seeing one for Qt for KDE-based Linux desktops.

To find out how to do all of these things, refer to
[Contributing Code to Toga](https://toga.readthedocs.io/en/stable/how-to/contribute/code.html)
in the documentation.
```