Exclude web from "issues not tagged with a specific platform"
There's a link "issues not tagged with a specific platform" in the "What to Do" section of the contribution docs that links to a GitHub page excluding all platform filters; however, that filter misses the Web backend, which we as a project treat as a platform.  There is no textual tag AFAIK so I'm not including that.